requirements.txt
asgiref==3.2.10
certifi==2023.7.22
charset-normalizer==3.3.1
cloudinary==1.30.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.6.0
djangorestframework==3.13.1
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.5
python-decouple==3.4
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==3.7.4.3
urllib3==1.26.18
whitenoise==6.5.0
